pyproject.toml
[tool.poetry]
name = "credproxy"
version = "0.1.0"
description = "Service to provide AWS IAM authentication via container endpoint URLs or credential files"
packages = [
    {include = "credproxy"}
]
authors = ["John Preston <john@ews-network.net>"]
license = "MPL-2.0"
classifiers=[
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)"
]
readme = "README.rst"
keywords=["aws", "iam", "containers", "secrets"]

[tool.poetry.dependencies]
python = "^3.11"
boto3 = ">=1.40,<2.0"
PyYAML = "^6.0.3"
jsonschema = "^4.25"
flask = "^3.1.2"
watchdog = "^4.0.0"
prometheus-client = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3"
tbump = "^6.9"
ruff = "^0.14.1"
doc8 = "^2.0.0"
cyclonedx-bom = "^7.2.1"

[tool.poetry.group.test.dependencies]
coverage = "^7.1"
pytest = "^8.4"
pytest-cov = "^7.0.0"

[tool.poetry.group.docs.dependencies]
sphinx = "^8.2"
sphinx-material = "^0.0.35"
sphinx-jsonschema = "^1.19.1"
furo = "^2025.9.25"
sphinx-sitemap = "^2.9.0"

[tool.poetry.group.tooling.dependencies]
cfn-flip = "^1.3.0"

[tool.poetry.scripts]
credproxy = "credproxy.cli:main"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
# Select rules to enable
select = [
    "E",   # pycodestyle errors (for E501 line length)
    "F",   # pyflakes (unused imports, undefined variables)
    "I",   # isort (import sorting)
    "UP",  # pyupgrade (Python syntax upgrades)
]
future-annotations = true

# Ignore specific rules
ignore = [
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
skip-magic-trailing-comma = false

[tool.isort]
profile = "black"
line_length = 88
case_sensitive = true
known_first_party = ["credproxy"]
combine_as_imports = true
split_on_trailing_comma = false
length_sort = true
lines_after_imports = 2

[tool.ruff.lint.isort]
# Match your isort settings
case-sensitive = true
known-first-party = ["credproxy"]
combine-as-imports = true
split-on-trailing-comma = false
length-sort = true
lines-after-imports = 2

[tool.coverage.run]
source = ["credproxy"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
    "credproxy/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "if TYPE_CHECKING:",
    "__git_commit__ = ",
    "__build_date__ = ",
    "__version__ = _build_version",
]

[tool.tbump]
github_url = "https://github.com/johnpreston/credproxy"

[tool.tbump.version]
current = "0.1.0"

regex = '''
  (?P<major>\d+)
  \.
  (?P<minor>\d+)
  \.
  (?P<patch>\d+)
  (?:(?P<rc>[\S]+))?
  '''

[tool.tbump.git]
message_template = "Bump to {new_version}"
tag_template = "v{new_version}"

# For each file to patch, add a [[tool.tbump.file]] config
# section containing the path of the file, relative to the
# tbump.toml location.
[[tool.tbump.file]]
src = "pyproject.toml"

[[tool.tbump.file]]
src = "credproxy/__init__.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
